added source

+++ java/com/captainhampton/android/copterclone/GameObject.java
@@ -5,7 +5,7 @@ import android.graphics.Rect;
 public abstract class GameObject {
     protected int x;
     protected int y;
-    protected int dy;
+    protected double dy;
     protected int dx;
     protected int width;
     protected int height;
